rename mod encode to vlq_encode;
