+++ sigma-ser/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod encode;
+pub mod vlq_encode;
 pub mod zig_zag_encode;
