+++ sigma-ser/src/vlq_encode.rs
